Cleaned up the API a little
