--- stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/SystemMessage.kt
@@ -1,28 +1,0 @@
-package io.getstream.chat.android.compose.ui.components.messages
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import io.getstream.chat.android.compose.state.messages.list.SystemMessageState
-import io.getstream.chat.android.compose.ui.theme.ChatTheme
-
-/**
- * A system message is a message generated by a system event, such as updating the channel or muting a user.
- *
- * @param systemMessageState The system message item to show.
- * @param modifier Modifier for styling.
- */
-@Composable
-public fun SystemMessage(
-    systemMessageState: SystemMessageState,
-    modifier: Modifier = Modifier,
-) {
-    Text(
-        modifier = modifier,
-        text = systemMessageState.message.text,
-        color = ChatTheme.colors.textLowEmphasis,
-        style = ChatTheme.typography.footnoteBold,
-        textAlign = TextAlign.Center
-    )
-}
